Extract `context` and `patch` to `assistant_context_editor`
