+++ crates/assistant/src/inline_assistant.rs
@@ -1,8 +1,9 @@
 use crate::{
     humanize_token_count, AssistantPanel, AssistantPanelEvent, CycleNextInlineAssist,
-    CyclePreviousInlineAssist, RequestType,
+    CyclePreviousInlineAssist,
 };
 use anyhow::{anyhow, Context as _, Result};
+use assistant_context_editor::RequestType;
 use assistant_settings::AssistantSettings;
 use client::{telemetry::Telemetry, ErrorExt};
 use collections::{hash_map, HashMap, HashSet, VecDeque};
