+++ crates/assistant_context_editor/Cargo.toml
@@ -10,3 +10,41 @@ workspace = true
 
 [lib]
 path = "src/assistant_context_editor.rs"
+
+[dependencies]
+anyhow.workspace = true
+assistant_slash_command.workspace = true
+assistant_slash_commands.workspace = true
+assistant_tool.workspace = true
+chrono.workspace = true
+client.workspace = true
+clock.workspace = true
+collections.workspace = true
+editor.workspace = true
+feature_flags.workspace = true
+fs.workspace = true
+futures.workspace = true
+gpui.workspace = true
+language.workspace = true
+language_model.workspace = true
+language_models.workspace = true
+log.workspace = true
+open_ai.workspace = true
+paths.workspace = true
+project.workspace = true
+prompt_library.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+smallvec.workspace = true
+smol.workspace = true
+strum.workspace = true
+telemetry_events.workspace = true
+text.workspace = true
+ui.workspace = true
+util.workspace = true
+uuid.workspace = true
+
+[dev-dependencies]
+pretty_assertions.workspace = true
+settings.workspace = true
+unindent.workspace = true
